+++ src/plonemeeting/restapi/tests/test_services_annexes.py
@@ -94,23 +94,32 @@ class testServiceAnnexes(BaseTestCase):
         annex_infos = response.json()[0]
         self.assertEqual(annex_infos['UID'], annex.UID())
 
-    def test_restapi_annexes_include_categorized_infos(self):
-        """@annexes, it is possible to get categorized element infos."""
+    def test_restapi_annexes_additional_values(self):
+        """@annexes, additional_values will actually
+        return categorized element infos."""
         self.changeUser("pmManager")
         item = self.create("MeetingItem")
         self.addAnnex(item)
         transaction.commit()
         item_url = item.absolute_url()
+        # category_id
         endpoint_url = "{0}/@annexes?fullobjects&include_all=false&" \
-            "include_categorized_infos=false".format(item_url)
+            "additional_values=category_id".format(item_url)
         response = self.api_session.get(endpoint_url)
         resp_json = response.json()
         self.assertFalse("category_title" in resp_json[0])
-        endpoint_url = "{0}/@annexes?fullobjects&include_all=true&" \
-            "include_categorized_infos=false".format(item_url)
+        self.assertEqual(resp_json[0]["category_id"], u'financial-analysis')
+        # *
+        endpoint_url = "{0}/@annexes?fullobjects&include_all=false&additional_values=*".format(
+            item_url)
         response = self.api_session.get(endpoint_url)
         resp_json = response.json()
-        self.assertEqual(resp_json[0]["category_title"], u'Financial analysis')
+        self.assertTrue("category_id" in resp_json[0])
+        self.assertTrue("category_title" in resp_json[0])
+        self.assertTrue("category_uid" in resp_json[0])
+        # there are some ignored values
+        self.assertFalse("allowedRolesAndUsers" in resp_json[0])
+        self.assertFalse("visible_for_groups" in resp_json[0])
 
 
 def test_suite():